arke-platform/requirements-google-drive.txt
google-api-python-client>=2.170.0
google-auth>=2.40.0
google-auth-oauthlib>=1.2.0
pypdf>=5.6.0
